+++ src/at_cmd.h
@@ -8,7 +8,7 @@
  * @copyright Copyright (c) 2021
  *
  */
-#include "WisBlock-API.h"
+#include "WisBlock-API-V2.h"
 
 #ifndef __AT_H__
 #define __AT_H__