+++ src/flash-esp32.cpp
@@ -10,7 +10,7 @@
  */
 #ifdef ESP32
 
-#include "WisBlock-API.h"
+#include "WisBlock-API-V2.h"
 
 #include <Preferences.h>