+++ src/flash-rp2040.cpp
@@ -10,7 +10,7 @@
  */
 #ifdef ARDUINO_ARCH_RP2040
 
-#include "WisBlock-API.h"
+#include "WisBlock-API-V2.h"
 
 s_lorawan_settings g_flash_content;
 s_loracompat_settings g_flash_content_compat;