+++ src/lora.cpp
@@ -8,7 +8,7 @@
  * @copyright Copyright (c) 2021
  *
  */
-#include "WisBlock-API.h"
+#include "WisBlock-API-V2.h"
 
 // LoRa callbacks
 static RadioEvents_t RadioEvents;